test: split RoslynTransformationTests into folder
